+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 ## Unreleased
 
+* Removed code that prints Baggage information
+  ([#1825](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1825))
+
+## 1.0.1
+
+Released 2021-Feb-10
+
+## 1.0.0-rc4
+
+Released 2021-Feb-09
+
+## 1.0.0-rc3
+
+Released 2021-Feb-04
+
+* Moved `ConsoleActivityExporter` and `ConsoleLogRecordExporter` classes to
+  `OpenTelemetry.Exporter` namespace.
+  ([#1770](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1770))
+
+## 1.0.0-rc2
+
+Released 2021-Jan-29
+
 * `AddConsoleExporter` extension method for logs moved from
   `OpenTelemetry.Trace` namespace to `OpenTelemetry.Logs` namespace.
   ([#1576](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1576))